% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_over_idents.R
\name{iterate_over_idents}
\alias{iterate_over_idents}
\title{Iterate over idents internal}
\usage{
iterate_over_idents(
  ident,
  keep.proportions,
  splits,
  cell.meta,
  subidents,
  idents,
  split.by,
  sample.sizes,
  equal_splits,
  seed = seed
)
}
\arguments{
\item{ident}{A character}

\item{keep.proportions}{A character}

\item{splits}{A character}

\item{cell.meta}{A character}

\item{subidents}{A character}

\item{idents}{A character}

\item{split.by}{A character}

\item{sample.sizes}{A character}

\item{equal_splits}{A character}

\item{seed}{A numeric}
}
\value{
Iteration over idents
}
\description{
Iterate over idents internal
}
