% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cells_hierarchically.R
\name{sample_cells_hierarchically}
\alias{sample_cells_hierarchically}
\title{sample_cells_hierarhically.R}
\usage{
sample_cells_hierarchically(
  cell.meta,
  idents,
  subidents,
  split.by,
  equal_splits = FALSE,
  as.data = FALSE,
  sample.size = NULL,
  keep.proportions = FALSE,
  seed = 42,
  plot = FALSE
)
}
\arguments{
\item{cell.meta}{A character}

\item{idents}{A character}

\item{subidents}{A character}

\item{split.by}{A character}

\item{equal_splits}{A character}

\item{as.data}{A character}

\item{sample.size}{A character}

\item{keep.proportions}{A character}

\item{seed}{A character}

\item{plot}{A character}
}
\value{
A numeric vector
}
\description{
sample_cells_hierarhically.R
}
\examples{
sample_cells_hierarchically()

}
