% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_by_name.R
\name{sort_by_name}
\alias{sort_by_name}
\title{sort_by_name}
\usage{
sort_by_name(x)
}
\arguments{
\item{x}{a named vector}
}
\value{
sorted vector
}
\description{
sort_by_name
}
\examples{
a=setNames(c(1,2,3), nm=c("C", "A", "B"))
sort_by_name(a)
}
