% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_split.R
\name{subset_split}
\alias{subset_split}
\title{Subset split internal}
\usage{
subset_split(
  split,
  cell.meta,
  subidents,
  idents,
  ident,
  split.by,
  sample.sizes,
  subtype.props,
  seed = seed
)
}
\arguments{
\item{split}{Split}

\item{cell.meta}{Data to subset}

\item{subidents}{A character}

\item{idents}{A character}

\item{ident}{A character}

\item{split.by}{Grouping variable}

\item{sample.sizes}{Numeric}

\item{subtype.props}{Logical}

\item{seed}{Numeric}
}
\value{
subset splits
}
\description{
Subset split internal
}
