% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test.R
\name{Test}
\alias{Test}
\title{Test.R}
\usage{
Test(message = "Test successfull")
}
\arguments{
\item{message}{A string to be returned as message}
}
\value{
Success message
}
\description{
Test.R
}
\examples{
Test()
Test("The test was successfull!")
}
